+++ mfhdf/hdfimport/hdfimporttest.c
@@ -170,16 +170,16 @@ main(void)
     sp = fopen("ctxtr2", "w");
     (void)fprintf(sp, "%s\n", text);
     (void)fprintf(sp, "%10d%10d%10d\n", ione, nrow, ncol);
-    (void)fprintf(sp, "%14.6E%14.6E\n", ezero, ezero);
+    (void)fprintf(sp, "%14.6E%14.6E\n", (double)ezero, (double)ezero);
     for (i = 0; i < nrow; i++)
-        (void)fprintf(sp, "%14.6E", row4[i]);
+        (void)fprintf(sp, "%14.6E", (double)row4[i]);
     (void)fprintf(sp, "\n");
     for (j = 0; j < ncol; j++)
-        (void)fprintf(sp, "%14.6E", col4[j]);
+        (void)fprintf(sp, "%14.6E", (double)col4[j]);
     (void)fprintf(sp, "\n");
     for (i = 0; i < nrow; i++) {
         for (j = 0; j < ncol; j++)
-            (void)fprintf(sp, "%14.6E", b32r2[i][j]);
+            (void)fprintf(sp, "%14.6E", (double)b32r2[i][j]);
         (void)fprintf(sp, "\n");
     }
     (void)fclose(sp);
@@ -238,20 +238,20 @@ main(void)
     sp = fopen("ctxtr3", "w");
     (void)fprintf(sp, "%s\n", text);
     (void)fprintf(sp, "%10d%10d%10d\n", npln, nrow, ncol);
-    (void)fprintf(sp, "%14.6E%14.6E\n", ezero, ezero);
+    (void)fprintf(sp, "%14.6E%14.6E\n", (double)ezero, (double)ezero);
     for (k = 0; k < npln; k++)
-        (void)fprintf(sp, "%14.6E", pln4[k]);
+        (void)fprintf(sp, "%14.6E", (double)pln4[k]);
     (void)fprintf(sp, "\n");
     for (i = 0; i < nrow; i++)
-        (void)fprintf(sp, "%14.6E", row4[i]);
+        (void)fprintf(sp, "%14.6E", (double)row4[i]);
     (void)fprintf(sp, "\n");
     for (j = 0; j < ncol; j++)
-        (void)fprintf(sp, "%14.6E", col4[j]);
+        (void)fprintf(sp, "%14.6E", (double)col4[j]);
     (void)fprintf(sp, "\n");
     for (k = 0; k < npln; k++)
         for (i = 0; i < nrow; i++) {
             for (j = 0; j < ncol; j++)
-                (void)fprintf(sp, "%14.6E", b32r3[k][i][j]);
+                (void)fprintf(sp, "%14.6E", (double)b32r3[k][i][j]);
             (void)fprintf(sp, "\n");
         }
     (void)fclose(sp);
